No bike pushing when foot=use_sidepath is set
The `foot=use_sidepath` tag is not properly documented, but is used over
30.000 times on ways in OSM. Similar to the properly documented and
accepted tag `bicycle=use_sidepath`, it indicates the pedestrians may
not use that road. The tag implies that a parallel road suitable for them
has been mapped.

Currently OSRM can route cyclists over busy intersections not suitable
for cyclists when these roads are tagged with `foot=use_sidepath`,
because this value is not seen as prohibiting dismounting and walking a
bit.

Dismounting cyclists should adhere to the rules for pedestrians. The
bicycle profile only checked for `foot=no`. This commit checks all
blacklisted access values instead.